requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==1.4.6
django-mysql==3.9
PyJWT == 2.3.0